Made a playable version
Changed scaling from int to float to provide .25 scale (=25%)
Added generation and progress of enemies.

+++ src/business/Cursor.java
@@ -5,7 +5,7 @@ import java.awt.Dimension;
 import java.awt.Graphics;
 
 public class Cursor extends GeometricRenderable {
-	public Cursor(int x, int y, int sizeX, int sizeY) {
+	public Cursor(int x, int y, float sizeX, float sizeY) {
 		super(x, y, sizeX, sizeY, ZINDEX_LAYER);
 	}
 	
@@ -13,18 +13,18 @@ public class Cursor extends GeometricRenderable {
 	{
 		// Draw a red circular crosshair
 		graphics.setColor(Color.RED);
-		graphics.drawOval(this.GetX(), this.GetY(), this.GetSizeX(), this.GetSizeY());
+		graphics.drawOval(this.GetX(), this.GetY(), (int)this.GetSizeX(), (int)this.GetSizeY());
 	}
 	
 	public void SetX(int x)
 	{
-		int maxWidth = GameLoop.Instance().GetSize().width - this.GetSizeX();
+		int maxWidth = (int)(GameLoop.Instance().GetSize().width - this.GetSizeX());
 		this.position.width = (x < 0 || x > maxWidth) ? (x < 0 ? 0 : maxWidth) : x;
 	}
 	
 	public void SetY(int y)
 	{
-		int maxHeight = GameLoop.Instance().GetSize().height - this.GetSizeY();
+		int maxHeight = (int)(GameLoop.Instance().GetSize().height - this.GetSizeY());
 		this.position.height = (y < 0 || y > maxHeight) ? (y < 0 ? 0 : maxHeight) : y;
 	}
