+++ src/business/ImageRenderable.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 public class ImageRenderable extends Renderable {
 	protected BufferedImage image;
 	
-	public ImageRenderable(int x, int y, int sizeX, int sizeY, int zIndex) {
+	public ImageRenderable(int x, int y, float sizeX, float sizeY, int zIndex) {
 		super(x, y, sizeX, sizeY, zIndex);
 	}
